refactor: move to tmp_path_builder.py to root directory
